+++ user-interface/client/app/src/components/website.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import {useState} from 'react';
 
+import '../stylesheets/website.css';
+
 import Banner from "./banner.js";
 import Page from "./page.js";
 
@@ -8,9 +10,11 @@ export default function Website() {
     const [page,setPage] = useState("home_page");
 
     return(
-            <div style={{display:"flex"}}>
-            <Banner page={page} toHomePage={() => setPage("home_page")} toDatabasePage={() => setPage("database_page")} toAboutPage={() => setPage("about_page")} toHelpPage={() => setPage("help_page")} toContactUsPage={() => setPage("contact_us_page")} toSettingsPage={() => setPage("settings_page")}/>
-            <Page page={page}/> 
+            <div id="overall_website">
+                <Banner page={page} toHomePage={() => setPage("home_page")} toDatabasePage={() => setPage("database_page")} toAboutPage={() => setPage("about_page")} toHelpPage={() => setPage("help_page")} toContactUsPage={() => setPage("contact_us_page")} toSettingsPage={() => setPage("settings_page")}/>
+                <div id="page_view">
+                    <Page page={page}/> 
+                </div>
             </div>
                 
         );